% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawNormalValues.R
\name{drawNormalValues}
\alias{drawNormalValues}
\title{drawNormalValues
Draws ui or ei values from N(groupMean, groupStandardDeviation).}
\usage{
drawNormalValues(sizeOfDraw, groupMean, groupStandardDeviation)
}
\arguments{
\item{sizeOfDraw}{The number of elements that need to be drawn.}

\item{groupMean}{The mean of the normal distribution used to draw ui or ei for a group.}

\item{groupStandardDeviation}{The standard deviation of the normal distributuion used to draw ui or ei for a group.}
}
\description{
drawNormalValues
Draws ui or ei values from N(groupMean, groupStandardDeviation).
}

