% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genVoterLevelDataFrame.R
\name{genVoterLevelDataFrame}
\alias{genVoterLevelDataFrame}
\title{genVoterLevelDataframe
Takes the information from a group level data frame generated by genGroupLevelDataframe(), and draws ui and ei for the groups of voters based on it.}
\usage{
genVoterLevelDataFrame(groupDataFrame)
}
\arguments{
\item{groupDataFrame:}{A data frame generated by genGroupLevelDataframe.}
}
\value{
voterDataFrame: A data frame with a row for each, containing the following columns: groupID, ui and ei.
}
\description{
genVoterLevelDataframe
Takes the information from a group level data frame generated by genGroupLevelDataframe(), and draws ui and ei for the groups of voters based on it.
}

