% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotDecisionCost.R
\name{plotDecisionCost}
\alias{plotDecisionCost}
\title{plotDecisionCost
Plots the mean number of rounds it took for the status quo to be defeated for each k-majority rule, across all the iterations.}
\usage{
plotDecisionCost(outputDataList)
}
\arguments{
\item{outputDataList}{The output data list of summaries, that is generated by the iterations() function}
}
\value{
A plot of the mean, across all iterations, of the decision cost incurred for each k-majority rule.
}
\description{
plotDecisionCost
Plots the mean number of rounds it took for the status quo to be defeated for each k-majority rule, across all the iterations.
}

