% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotExpectedUtilityTotalCost.R
\name{plotExpectedUtilityTotalCost}
\alias{plotExpectedUtilityTotalCost}
\title{plotExpectedUtilityTotalCost}
\usage{
plotExpectedUtilityTotalCost(outputDataList, plotMeanEU, plotBestEU,
  plotWorstEU)
}
\arguments{
\item{outputDataList:}{The output data list of summaries, that is generated by the iterations() function}

\item{plotMeanEU:}{if TRUE plot the costs for the Mean Group}

\item{plotBestEU:}{if TRUE plot the costs for the Best Group}

\item{plotWorstEU:}{if TRUE plot the costs for the Worst Group}
}
\value{
A plot of the mean, across all iterations, of the decision cost, expected utility and total cost incurred for each k-majority rule.
}
\description{
Plots the Total Cost, Decision Cost and Expected Utility measures for the voters of interest, in the round the proposal passed, for each of the kMajority rules.
}

