% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotNumberOfRounds.R
\name{plotNumberOfRounds}
\alias{plotNumberOfRounds}
\title{plotNumberOfRounds
Plots the mean number of rounds it took for the status quo to be defeated for each k-majority rule, across all the iterations.}
\usage{
plotNumberOfRounds(outputDataList)
}
\arguments{
\item{outputDataList}{The output data list of summaries, that is generated by the iterations() function}
}
\value{
A plot of the mean, across all iterations, that a proposal passed for each k-majority rule.
}
\description{
plotNumberOfRounds
Plots the mean number of rounds it took for the status quo to be defeated for each k-majority rule, across all the iterations.
}

