% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotOnlyExpectedUtility.R
\name{plotOnlyExpectedUtility}
\alias{plotOnlyExpectedUtility}
\title{plotOnlyExpectedUtilityCost}
\usage{
plotOnlyExpectedUtility(outputDataList)
}
\arguments{
\item{outputDataList:}{The output data list of summaries, that is generated by the iterations() function}
}
\value{
A plot of the mean, across all iterations, of the external cost incurred for each k-majority rule.
}
\description{
Plots the expected utility measures for the voters of interest, in the round the proposal passed, for each of the kMajority rules.
}

