% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPareto.R
\name{plotPareto}
\alias{plotPareto}
\title{plotPareto}
\usage{
plotPareto(outputDataList)
}
\arguments{
\item{outputDataList}{The output data list of summaries, that is generated by the iterations() function}
}
\value{
A plot of the umber of rounds (y-axis) where a Pareto preferred proposal failed by k-majority (x-axis).
}
\description{
Plot the number of rounds (y-axis) where a Pareto preferred proposal failed by k-majority (x-axis)
}

